error renderer, connections, sync

+++ src/formBuilder/elements/field.js
@@ -1,21 +1,29 @@
-import {fieldSupportedListenerTypes, fieldValueAccessor} from '../shared/constants';
-import {noop} from "../utils";
-import Capitalize from '../utils/capitalize'
-import Compose from '../utils/compose'
+import {fieldSupportedListenerTypes, fieldValueAccessor} from "../shared/constants";
+import {capitalize, compose, noop} from "../utils";
 
 const defaultConfiguration = {
     validators: [],
     validClassName: 'valid',
     invalidClassName: 'invalid',
     errorContainerSelector: '.errors',
-    onError: noop,
-    onValid: noop,
     errorMessageFormatter: msg => `<p style="color: red;">${msg}</p>`,
 };
 
+const defaultHooks = {
+    input: noop,
+    change: noop,
+    error: noop,
+    blur: noop,
+    focus: noop,
+    reset: noop,
+    destroy: noop,
+    valid: noop,
+    init: noop,
+};
+
 class Field {
-    constructor({name, configuration}) {
-        this.configuration = Object.assign({}, defaultConfiguration, configuration);
+    constructor({name, configuration, value}) {
+        this.configuration = {...defaultConfiguration, ...configuration};
         this.name = name;
         this.fieldNode = document.querySelector(`[name="${name}"]`);
 
@@ -23,6 +31,7 @@ class Field {
             throw new Error('Field element does\'not exist');
         }
 
+        this.hooks = {...defaultHooks, ...configuration.hooks};
         this.type = this.fieldNode.getAttribute('type') || 'text';
         this.valueAccessor = fieldValueAccessor[this.type];
         this.validators = this.configuration.validators;
@@ -30,14 +39,20 @@ class Field {
         this.dirty = null;
         this.touched = null;
         this._value = this.fieldNode[this.valueAccessor];
-        this.defaultValue = this.value;
+        this.defaultValue = value || this.value;
         this.invalid = null;
         this.parent = null;
         this.focus = false;
         this.errors = [];
-        this.middleware = Compose.apply(null, this.configuration.middlewares);
+        this.pipes = (compose.apply(null, this.configuration.pipes)).bind(this);
         this.errorContainer = this.fieldNode.parentElement.querySelector(this.configuration.errorContainerSelector);
         this.attachListeners();
+
+        if (value) {
+            this.value = value;
+        }
+
+        this.hooks.init(this);
     }
 
     get value() {
@@ -51,16 +66,25 @@ class Field {
 
     setParent(parent) {
         this.parent = parent;
+        this.hooks.change(null, this);
     }
 
     attachListeners() {
         if (!fieldSupportedListenerTypes[this.type]) return;
         for (let listenerType of fieldSupportedListenerTypes[this.type]) {
-            const eventMethod = 'on' + Capitalize(listenerType);
+            const eventMethod = 'on' + capitalize(listenerType);
             if (!this[eventMethod]) continue;
-            this.fieldNode.addEventListener(listenerType, event => {
-                this[eventMethod](event);
-            });
+            this[eventMethod] = this[eventMethod].bind(this);
+            this.fieldNode.addEventListener(listenerType, this[eventMethod], true);
+        }
+    }
+
+    detachListeners() {
+        if (!fieldSupportedListenerTypes[this.type]) return;
+        for (let listenerType of fieldSupportedListenerTypes[this.type]) {
+            const eventMethod = 'on' + capitalize(listenerType);
+            if (!this[eventMethod]) continue;
+            this.fieldNode.removeEventListener(listenerType, this[eventMethod], true);
         }
     }
 
@@ -77,24 +101,23 @@ class Field {
             }
         }
 
-        if (this.invalid) {
-            this.renderErrors();
-        } else {
-            this.clearErrors();
-        }
+        this.invalid ? this.renderErrors() : this.clearErrors();
 
         if (this.parent) {
             this.parent.validate();
         }
 
-        this.fieldNode.classList[this.valid ? 'add' : 'remove'](this.configuration.validClassName);
-        this.fieldNode.classList[this.invalid ? 'add' : 'remove'](this.configuration.invalidClassName);
+        if (this.validators.length > 0) {
+            this.fieldNode.classList[this.valid ? 'add' : 'remove'](this.configuration.validClassName);
+            this.fieldNode.classList[this.invalid ? 'add' : 'remove'](this.configuration.invalidClassName);
+        }
+
         this.valid ? this.onValid() : this.onError();
     }
 
-    runMiddleware(value) {
-        if (this.middleware) {
-            return this.middleware.call(this, value);
+    runPipes(value) {
+        if (this.pipes) {
+            return this.pipes(value);
         }
 
         return value;
@@ -103,27 +126,32 @@ class Field {
     onChange(event) {
         this.dirty = true;
         const oldValue = this.value;
-        const newValue = this.runMiddleware(this.fieldNode[this.valueAccessor]);
+        const newValue = this.runPipes(this.fieldNode[this.valueAccessor]);
         this.value = newValue;
 
         if (oldValue !== newValue) {
             this.validate();
         }
+
+        this.hooks.change(event, this);
     }
 
     onFocus(event) {
         this.touched = true;
         this.focus = true;
+        this.hooks.focus(event, this);
     }
 
-    onBlur() {
+    onBlur(event) {
         this.focus = false;
+        this.hooks.blur(event, this);
     }
 
     onInput(event) {
         this.dirty = true;
-        this.value = this.runMiddleware(this.fieldNode[this.valueAccessor]);
+        this.value = this.runPipes(this.fieldNode[this.valueAccessor]);
         this.validate();
+        this.hooks.input(event, this);
     }
 
     renderErrors() {
@@ -136,20 +164,23 @@ class Field {
         this.errorContainer.innerHTML = '';
     }
 
-    onError() {
-        this.configuration.onError.call(this);
+    onError(event) {
+        this.hooks.error(event, this);
     }
 
     onValid() {
-        this.configuration.onValid.call(this);
+        this.hooks.valid(this);
     }
 
     destroy() {
         this.fieldNode.remove();
+        this.detachListeners();
 
-        if (this.parent) {
+        if (this.parent && this.parent.destroyField) {
             this.parent.destroyField(this);
         }
+
+        this.hooks.destroy();
     }
 
     reset() {
@@ -160,6 +191,7 @@ class Field {
         this.invalid = null;
         this.touched = null;
         this.dirty = null;
+        this.hooks.reset(this);
     }
 }
